backend/requirements.txt
Django==6.0.3
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
django-cors-headers==4.9.0
django-filter==25.1
python-dotenv==1.2.2
mysqlclient==2.2.8
Pillow==10.2.0
celery==5.4.0
redis==5.0.1
gunicorn==23.0.0
